Added Sparql code to dedicated file
