+++ nvsvocprez/.env.example
@@ -3,3 +3,4 @@ AUTH0_CLIENT_SECRET=
 AUTH0_DOMAIN=
 APP_SECRET_KEY=
 OAUTH_BASE_URL=
+LOGIN_ENABLE=
